requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.5.0
    # via django
backports-zoneinfo==0.2.1
    # via django
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==2.0.12
    # via requests
colorlog==6.6.0
    # via -r requirements.in
cryptography==36.0.2
    # via -r requirements.in
deprecated==1.2.13
    # via pygithub
django==4.0.3
    # via
    #   -r requirements.in
    #   django-extensions
django-extensions==3.1.5
    # via -r requirements.in
idna==3.3
    # via requests
pycparser==2.21
    # via cffi
pygithub==1.55
    # via -r requirements.in
pyjwt==2.3.0
    # via
    #   -r requirements.in
    #   pygithub
pynacl==1.5.0
    # via pygithub
python-json-logger==2.0.2
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   pygithub
sqlparse==0.4.2
    # via django
urllib3==1.26.9
    # via requests
wrapt==1.14.0
    # via deprecated
